Split EndpointHtmlRenderer into themed partial class files
